delete-transaction-orders

+++ app/Http/Controllers/Frontend/ShoppingCartController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Frontend;
 use App\Http\Controllers\Controller;
 use Gloudemans\Shoppingcart\Facades\Cart;
 use App\Models\Transaction;
+use App\Models\User;
 use App\Models\Order;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
@@ -23,16 +24,20 @@ class ShoppingCartController extends Controller
     public function checkout()
     {
         $products = Cart::content();
+        $user = User::find(get_data_user('web'));
+        
         if(Cart::count() < 1)
         {
             $viewData = [
-                'message' => 'Giỏ hàng trống',
+                'user'     => $user,
+                'message'  => 'Giỏ hàng trống',
                 'products' => $products
             ];
             return view('frontend.shopping.index', $viewData);
         } else {
             
             $viewData = [
+                'user'     => $user,
                 'products' => $products
             ];
         return view('frontend.shopping.checkout', $viewData);
@@ -41,27 +46,32 @@ class ShoppingCartController extends Controller
 
     public function pay(Request $request)
     {
-        $dataTransaction                  = $request->except('_token');
-        $dataTransaction['created_at']    = Carbon::now();
-        $dataTransaction['t_total_money'] = (int)str_replace(',', '', Cart::subtotal(0));
-        $transaction = Transaction::create($dataTransaction);
-    
-        if($transaction)
+        if(Cart::count() > 0)
         {
-            $products = Cart::content();
-            foreach($products as $item)
+            $dataTransaction                  = $request->except('_token');
+            $dataTransaction['created_at']    = Carbon::now();
+            $dataTransaction['t_user_id']     = get_data_user('web') ?? 0;
+            $dataTransaction['t_total_money'] = (int)str_replace(',', '', Cart::subtotal(0));
+            $transaction = Transaction::create($dataTransaction);
+        
+            if($transaction)
             {
-                Order::create([
-                    'od_transaction_id' => $transaction->id,
-                    'od_product_id'     => $item->id,
-                    'od_qty'            => $item->qty,
-                    'od_price'          => $item->price,
-                    'created_at'        => Carbon::now()
-                ]);
+                $products = Cart::content();
+                foreach($products as $item)
+                {
+                    Order::create([
+                        'od_transaction_id' => $transaction->id,
+                        'od_product_id'     => $item->id,
+                        'od_qty'            => $item->qty,
+                        'od_price'          => $item->price,
+                        'created_at'        => Carbon::now()
+                    ]);
+                }
             }
-        }
 
-        Cart::destroy();
+            Cart::destroy();
+            return redirect()->route('get.home');
+        }
         return redirect()->route('get.home');
     }
 }
